refactor: Application layer
