+++ CruiserJumpPractice/Application/Services/ClientCruiserStateService.cs
@@ -2,9 +2,9 @@
 
 using BepInEx.Logging;
 using CruiserJumpPractice.Domain;
-using CruiserJumpPractice.UseCases;
+using CruiserJumpPractice.Application.UseCases;
 
-namespace CruiserJumpPractice.Services;
+namespace CruiserJumpPractice.Application.Services;
 
 internal sealed class ClientCruiserStateService
 {
